requirements.txt
requests>=2.32.3
beautifulsoup4>=4.12.3
pytest>=8.3.2
python-dotenv>=1.0.1
streamlit>=1.36.0
